fix: audit batches 209-215 — 9 files fixed across 34 audited (7606+6171+7083+6157+5318+4865+4816L)
Batch 209 (7606L, 5 files): plan/rewrite.rs 1M+1L (dead-node rule application,
  join dedup removes only first shared child), record/task.rs SOUND,
  observability/otel.rs SOUND, monitor.rs SOUND, console.rs SOUND

Batch 210 (6171L, 5 files): time/intrusive_wheel.rs 1M (u64→usize truncation
  on 32-bit in advance_to), trace/replay.rs SOUND, error.rs SOUND,
  combinator/race.rs 1M (loser panic overrides winner success), time/sleep.rs SOUND

Batch 211 (7083L, 6 files): io/browser_storage.rs SOUND, web/extract.rs SOUND,
  trace/canonicalize.rs SOUND, trace/distributed/vclock.rs SOUND,
  web/session.rs 1H (CRLF injection in Set-Cookie via cookie_name/path),
  time/interval.rs SOUND

Batch 212 (6157L, 6 files): observability/task_inspector.rs SOUND,
  analyzer_plugin.rs SOUND, combinator/join.rs SOUND, distributed/tests.rs SOUND,
  messaging/kafka_consumer.rs 1M (TOCTOU race in subscribe/poll/commit_offsets),
  obligation/ledger.rs SOUND

Batch 213 (5318L, 6 files): lab/network/network.rs SOUND, io/ext/read_ext.rs SOUND,
  observability/resource_accounting.rs 2L (overflow in pending/rejection_rate),
  encoding.rs 1M+2L (as u32 truncation, k+repair overflow),
  combinator/map_reduce.rs SOUND, web/router.rs SOUND

Batch 214 (4865L, 6 files): ALL SOUND — record/obligation.rs, tls/types.rs,
  trace/refinement_firewall.rs, combinator/first_ok.rs, select.rs, timeout.rs

Batch 215 (4816L, 6 files): lab/meta/mutation.rs SOUND, obligation/guarded.rs SOUND,
  lab/injection.rs 1M (lab-only unbounded poll loop, not fixed),
  net/websocket/handshake.rs 2M+3L (CRLF injection in headers/protocols/reject),
  net/unix/stream.rs SOUND, trace/compat.rs 2M+1L (unbounded alloc from untrusted length)

+++ src/time/intrusive_wheel.rs
@@ -494,7 +494,7 @@ impl<const SLOTS: usize> TimerWheel<SLOTS> {
         let ticks_to_advance = target_tick_u64 - self.current_tick;
 
         // If advancing more than SLOTS ticks, we need to scan all slots
-        if ticks_to_advance as usize >= SLOTS {
+        if ticks_to_advance >= SLOTS as u64 {
             // Full rotation or more: collect expired from ALL slots
             for slot in &self.slots {
                 let wakers = slot.collect_expired(now);
